# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Akamai CCM API
  version: 1.0.0
  description: >
    Cloud Compliance Manager API for managing Linode accounts, resources,
    compliance rules, reports, and users.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://<your-host>
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ComplianceRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        resource_types:
          type: array
          items:
            type: string
        condition_type:
          type: string
        condition_config:
          type: object
        is_overridden_disabled:
          type: boolean
        account_id:
          type: string
          format: uuid
          nullable: true
    ComplianceResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rule_id:
          type: string
          format: uuid
        resource_id:
          type: string
          format: uuid
          nullable: true
        account_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [compliant, non_compliant, not_applicable]
        detail:
          type: string
        acknowledged:
          type: boolean
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        acknowledged_note:
          type: string
          nullable: true
        acknowledged_by:
          type: string
          format: uuid
          nullable: true
        acknowledged_by_name:
          type: string
          nullable: true
        evaluated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        rule:
          $ref: '#/components/schemas/ComplianceRule'
        resource:
          $ref: '#/components/schemas/Resource'
          nullable: true
    ComplianceNote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        compliance_result_id:
          type: string
          format: uuid
        note:
          type: string
        created_at:
          type: string
          format: date-time
        author_name:
          type: string
          nullable: true
...
    ComplianceScoreHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        evaluated_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /api/compliance/results:
    get:
      tags: [Compliance]
      summary: List compliance results
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [compliant, non_compliant, not_applicable]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, warning, info]
        - name: resource_type
          in: query
          schema:
            type: string
        - name: rule_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of compliance results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceResult'
  /api/compliance/results/bulk-acknowledge:
    put:
      tags: [Compliance]
      summary: Bulk acknowledge compliance results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [result_ids, acknowledged]
              properties:
                result_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                acknowledged:
                  type: boolean
                acknowledged_note:
                  type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  updated:
                    type: integer
  /api/compliance/results/{result_id}/acknowledge:
    parameters:
      - name: result_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Compliance]
      summary: Acknowledge a compliance result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [acknowledged]
              properties:
                acknowledged:
                  type: boolean
                acknowledged_note:
                  type: string
      responses:
        '200':
          description: Updated acknowledgement
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  acknowledged:
                    type: boolean
                  acknowledged_at:
                    type: string
                    format: date-time
                    nullable: true
                  acknowledged_note:
                    type: string
                    nullable: true
  /api/compliance/results/{result_id}/notes:
    parameters:
      - name: result_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Compliance]
      summary: Get notes for a compliance result
      responses:
        '200':
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceNote'
    post:
      tags: [Compliance]
      summary: Add note to a compliance result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note:
                  type: string
      responses:
        '200':
          description: Created note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceNote'
  /api/compliance/score:
    get:
      tags: [Compliance]
      summary: Get current compliance score
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Latest compliance score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceScoreHistory'
  /api/compliance/score/history:
    get:
      tags: [Compliance]
      summary: Get compliance score history
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Score history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceScoreHistory'
  /api/compliance/resources/{resource_id}/timeline:
    parameters:
      - name: resource_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Compliance]
      summary: Get compliance timeline for a resource
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resource compliance timeline
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshots:
                    type: array
                    items:
                      type: object
                      properties:
                        snapshot_time:
                          type: string
                          format: date-time
                        compliant:
                          type: integer
                        non_compliant:
                          type: integer
                        not_applicable:
                          type: integer
                        total:
                          type: integer
                        compliance_score:
                          type: number
                          format: float
                  rule_history:
                    type: array
                    items:
                      type: object
                      properties:
                        rule_id:
                          type: string
                          format: uuid
                        rule_name:
                          type: string
                        severity:
                          type: string
                        resource_types:
                          type: array
                          items:
                            type: string
                        snapshot_time:
                          type: string
                          format: date-time
                        status:
                          type: string
                        detail:
                          type: string
  /api/compliance/rules:
    get:
      tags: [Compliance]
      summary: List compliance rules
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
          description: If provided, includes account-specific overrides
      responses:
        '200':
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceRule'
  /api/compliance/rules/overrides:
    get:
      tags: [Compliance]
      summary: Get rule overrides for an account
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of rule overrides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleOverride'
  /api/compliance/rules/{rule_id}/override:
    parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Compliance]
      summary: Set rule override
      description: Enable or disable a rule for an account. Requires `admin` role.
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active:
                  type: boolean
      responses:
        '200':
          description: Override applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  rule_id:
                    type: string
                    format: uuid
                  is_active:
                    type: boolean
  /api/compliance/rules/{rule_id}/config:
    parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Compliance]
      summary: Get rule config override
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rule configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfig'
    put:
      tags: [Compliance]
      summary: Set rule config override
      description: Requires `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, config_override]
              properties:
                account_id:
                  type: string
                  format: uuid
                config_override:
                  type: object
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfig'
...
  /api/compliance/profiles/active:
    get:
      tags: [Compliance]
      summary: Get active profiles for an account
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceProfile'
    put:
      tags: [Compliance]
      summary: Set active profiles for an account
      description: Requires `admin` role.
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile_ids]
              properties:
                profile_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Profiles updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  profile_ids:
                    type: array
                    items:
                      type: string
                      format: uuid
  /api/compliance/profiles/{profile_id}/activate:
    parameters:
      - name: profile_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Compliance]
      summary: Activate or deactivate a profile
      description: Requires `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, active]
              properties:
                account_id:
                  type: string
                  format: uuid
                active:
                  type: boolean
      responses:
        '200':
          description: Profile activation updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean

  # ─── Reports ──────────────────────────────────────────────────────────────
...